pasarela de pago completa

+++ app/src/main/java/com/example/elretornodeloretro/model/TipePay.kt
@@ -3,5 +3,5 @@ package com.example.elretornodeloretro.model
 data class TipePay(
     val DESCRIPCION: String,
     val FORMA_DE_PAGO: String,
-    val ID_PAGO: Float
+    val ID_PAGO: Int
 )
